Add files via upload
first upload of arduino code.
The sensor logs show the temprature drift of Rgas when the Adafruit libraries are used. The Adafruit libraries rely on BME680.c written by Bosch. The temprature drift is too big for any usefull usage of the BME680 in this project. To use the sensor the closed source library from Bosch is needed. Initially this would not compile but changes to platform.txt make this possible. The documentation on platform.txt by Bosch is faulty.
The libraries and platform.txt will be added later, after I read their license agreement.